chore: Make mentor page layout
