+++ app/(dashboard)/(routes)/admin/messages/page.tsx
@@ -1,9 +1,9 @@
-import Main from "@/components/shared/Main";
+import Main from "@/components/shared/layout/mentor-main";
 
 export default function Default() {
   return (
     <>
-       <Main page="mentor"/>
+      <Main page="mentor" />
     </>
   );
 }
