+++ components/views/User.tsx
@@ -1,11 +1,11 @@
 
-import Main from '@/components/shared/Main'
+import Main from '@/components/shared/layout/mentor-main'
 
 const UserView = () => {
   return (
     <div className="flex flex-col w-full h-screen">
       {/* <Header page="user" /> */}
-      <Main page="user"/>
+      <Main page="user" />
     </div>
   )
 }
