+++ types/admin.ts
@@ -1,0 +1,3 @@
+export interface DashboardLayoutProps {
+  children: React.ReactNode;
+}
