+++ types/index.d.ts
@@ -2,6 +2,10 @@ declare global {
   interface Window {
 
   }
+
+  interface HeaderProps {
+    title: string;
+  }
 }
 
 export {};
